requirements.txt
flask
openai-whisper
gtts
torch